refactor: 페이지 이름 변경
- CompetitionCreatePage -> CreateCompetitionPage (영문법에 맞춰 수정함)
