+++ frontend/src/pages/CreateCompetitionPage.tsx
